Initial version of polar embeddings (#76)
* Initial version of polar embeddings

Couple of issues:

This is a stateful component, but it doesn't train. So maybe it
shouldn't be a tutorial project, but somewhere else?

I used the IMDB reviews dataset from Kaggle for this but it looks like
there's no public URL to download the same version of the data. The
upstream URL provided at Kaggle has the same data, but every tiny review
gets its own file so the overhead takes it from 60MB to 480MB on disk.
The dataset is too big anyway, so maybe there's a better alternative?

To calculate polar axes properly, the reference words should be averaged
with the nearest words in the vocab, but I'm not doing that yet.

* Get average of neighbor vectors

* Change line width

* Apply suggestions from code review

Co-authored-by: Sofie Van Landeghem <svlandeg@users.noreply.github.com>

* Move to pipelines

* Remove spurious comment

* Add README, update assets stuff, expand description

* Update README

Co-authored-by: Sofie Van Landeghem <svlandeg@users.noreply.github.com>