Add data augmentation tutorial (#77)
* Setup project skeleton

* Add script for preprocessing the file

* Add raw.txt to gitignore

No need to commit this into the repository because they can just
download that from the repository itself.

* Add requirements.txt

* Add black-compatible flake8 config

* Use a model immediately to include the entities

* Add data augmentation step-by-step notebook

* Remove raw.txt from .gitignore

* Update preprocess commands for training and test

* Remove dependency on combined raw.spacy

* Add raw training data and test set

Here, I added the following:

train_initial.jsonl: will be converted later to train.spacy and
dev.spacy after augmentation. This will be the raw text that needs to be
augmented further.

test.jsonl: the held-out test to evaluate our models against. Will be
converted into .spacy format in the preprocess step.

* Add directories for the input data

* Simplify the preprocessing step further

* Add directory to store metrics

* Add docstrings for preprocess.py

* Update project.yml with valid directories

* Implement data augmentation module

I decided to separate the actual weak supervision module from the one
that the CLI calls. For more control, I implemented flags so that one
can decide if they want weak supervision OR nlpaug.

* Update project.yml to include augment command

* Add explicit statements for output assets and deps

* Add initial configuration

* Ignore training directory in git

* Improve documentation and fix a few bugs

* Remove unnecessary return values

* Fix incorrect paths in project.yml

* Add cells for testing your module

* Fix bug on return value

* Update config to use initialize vectors

* Remove wiki_uncased gazetteers

It affects training accuracy because it predicts unreliable spans.

* Format project.yml properly for readability

* Remove .gitkeep from generated directories

Because these directories are automatically created when running
spacy <command> there's no need for it to exist prior.

* Add all generated folders in .gitignore

* Streamline .gitignore

We already have a generic gitignore in the root of explosion/projects
We can just have a project-speficic one here.

* Add command for decompressing zipfiles

* Update I/O in project.yml file

* Add auto-generated README

* Update category docs via script

* Ignore jupyter notebooks

* Add README and quick explanation of data

* Whitelist data README and demo to gitignore

* Add package step

Don't add this to any of the workflows first, I'd rather do the publish
/ package step manually for now.

* Cleanup labels and remove ORGS

We only check for the PERSON entity so let's just remove the ORG
entity. We don't need it for reporting.

* Move config.cfg into its own directory

* Add workflow for using spaCy builtin augmenter

* Add orth variants file in assets

* Add dependencies based on contextual substitution

* Update README to include builtin augmenter

* Remove other tokens to keep README minimal

* Fix links in assets README

* Fix docstring errors

* Update constants to include the source

* Include weak supervision and heuristic explainer

* Fix some grammar on README

* Apply suggestions from code review

- Updated docstrings
- Corrected a few paths

Co-authored-by: Sofie Van Landeghem <svlandeg@users.noreply.github.com>

* Use only one config file

Instead of having two configuration files, we just have one that
contains the augmenter itself. To deactivate it, we just pass "null"
to `corpora.train.augmenter`.

* Rename default to spacy-no-augmenter for explicitness

* Remove torch, sentencepiece, and transformers from deps

* Add install command

This command installs all dependencies from requirements.txt

* Apply suggestions from code review

Co-authored-by: Sofie Van Landeghem <svlandeg@users.noreply.github.com>

* Rename train and test for explicitness

* Move longer description in README to project.yml

Ideally, we shouldn't edit the README itself. It should just be an
artifact of whatever's inside our project.yml

* Regenerate README with new configuration

* Rename eval_* to dev_*

* Rename asset files from directory

* Add a limit command for faster testing

* Revert deletion of config_no_augmenter.cfg

* Add integration test for end-to-end run

* Include train-with-augmenter in the workflow

* Add the --user flag during pip install

Windows build is failing because of some permissions.

* Expand explanation of weak supervision

* Bump version to v1.0.0

* Add fixes on formatting

Co-authored-by: Sofie Van Landeghem <svlandeg@users.noreply.github.com>