Short circuit when preparing span data
The script to build span data is comparatively slow because it doesn't
assume that span groups are necessarily aligned. Because of that it
walks through all possible spans and finds the smallest containing span
to use as the gold annotation.

This change short-circuits when a containing span of size one is found.
With this change processing time is around 6m, where previously it was
(based on recollection) around 20m.