Rewrite FIXME
The note in the FIXME suggests the script the script being slow because
it checks everything is a problem. However, from the perspective of
correctness, checking everything is necessary for two reasons:

1. Given this script uses non-gold word-level spans, there's no guarantee
   of spangroup alignment.
2. Even with spangroup alignment, there's no guarantee a head has a
   unique enclosing span. (Logically it would normally have a single
   enclosing span, but as a data format issue it's capable of
   representing different enclosing spans.)

Because the only input to the span resolver is the token, it can't
attempt to learn different correct enclosing spans for a single token,
which is why we need to take one (in this case, the shortest) overall.

If this script were slower it might be worth sacrificing some
correctness for speed, but with the short circuit it's fast enough for
now.