Add use_gold_heads option
In generating training data for the span resolver we use predicted
heads instead of gold heads, but it would also be reasonable to just use
the gold heads. This adds the option to do that. Unfortunately, it
didn't have any noticeable effect on accuracy.

This also fixes an indentation issue that caused blank clusters to end
up in training docs sometimes. They were just ignored, so it didn't
actually affect training in any way.