Add separate configs to run test with tok2vec
Debugging suggests the issue in the test env is memory usage. This
commit makes several changes to allow running tests with a tok2vec,
instead of a Transformer, to address that.

1. add test configs
2. modify prep-span-data
3. add configs path to project.yml and use it in various steps

If this works as intended, it would probably be best to make the config
structure parallel for trf and tok2vec pipelines. However, since it's
unlikely it's possible to get good performance with a tok2vec, it may
make sense to emphasize that the configs are specifically for testing,
unlike the tok2vec configs in other projects.