Add span cleaner to output pipeline
The span cleaner is a new component that strips spans based on a key
prefix. In this case it will be used to remove the head spans. This is
implemented as a component so that it can be trivially disabled at
runtime in case it's necessary for debugging, or just because heads are
easier to work with for some reason.