Change prep-span-data to use silver, +cleanup
This changes the prep-span-data script to use silver heads by default.
This doesn't make training faster or more accurate, but it avoids
warnings caused by misalignment that come from using gold heads. There's
also structural changes to the script to make it easier to specify this
option in the project file.

This also fixes an issue with specifying gpu support in the scripts.