Use brute-force for feature ablation experiment
Unfortunately, string interpolation in spacy projects (when using list
of strings) is janky. TODO: fix it. In the meantime, I created multiple
configs for each ablation combination, then call it in the
feature-ablation command. I will port it to the more efficient version
once I've fixed this issue in Thinc.