Add a script for merging pipelines
This project includes a script to merge pipelines, handling issues like
whether to use replace_listeners and giving warnings when
replace_listeners isn't suitable.

Currently this should work well if a secondary pipeline has only one
listener, or if it's ok that a merged pipeline can't be trained.
However, there are still many things that can be improved, and questions
about how to do things like ensure compatibility between the pipelines
to be merged.

I'm also not sure a project is the best structure for this - it might
also make sense as a spaCy command. This seemed to be the best way to
share it for discussion though.