Add creation of projects.jsonl in docs update script (#150)
* Add step to create projects.jsonl

This adds the creation of a projects.jsonl file to the docs update
script. The resulting file can be used to get a list of projects in the
repo.

* Add initial version of built projects.jsonl

* Create separate script to update projects.jsonl

* Move details from coref project.yml to README

* Move detail to README from project.yml

* Add note on json script to top-level README

* Move floret details to README file

This also removes a duplicate section about wiki-extractor that was
repeated for some reason.

* Regenerate floret README

* Update weak supervision README

* Add back deleted READMEs

For the floret wiki and ner tweet projects, the part of the docs deleted
from project.yml wasn't actually added to the READMEs. This remedies
that.